# docs/swagger.yml
basePath: /api/v1
consumes:
    - application/json
definitions:
    BaseResponseBody:
        description: Contains error and message fields.
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
        required:
            - error
            - message
        title: BaseResponseBody represents the base response body for API.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    Comment:
        description: Comment represents the comment for this application
        properties:
            content:
                description: Comment content
                type: string
                x-go-name: Content
            created_at:
                description: The time the comment was created
                format: date-time
                type: string
                x-go-name: CreatedAt
            id:
                description: The id for this comment
                format: uuid
                type: string
                x-go-name: ID
            liked_by_me:
                type: boolean
                x-go-name: LikedByMe
            likes:
                description: Number of likes
                format: int64
                type: integer
                x-go-name: Likes
            updated_at:
                description: The time the comment was updated
                format: date-time
                type: string
                x-go-name: UpdatedAt
            user:
                $ref: '#/definitions/User'
        required:
            - id
            - content
            - created_at
            - updated_at
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/models
...
    GetPostResponseBody:
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            post:
                $ref: '#/definitions/Post'
        required:
            - error
            - message
            - post
        title: GetPostResponseBody includes the signle post for a given ID.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    GetPostsResponseBody:
        properties:
            count:
                format: int64
                type: integer
                x-go-name: Count
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            posts:
                items:
                    $ref: '#/definitions/Post'
                type: array
                x-go-name: Posts
        required:
            - error
            - message
            - count
            - posts
        title: GetPostsResponseBody includes the slice of posts.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
...
    Post:
        description: Post represents the post for this application
        properties:
            comments:
                items:
                    $ref: '#/definitions/Comment'
                type: array
                x-go-name: Comments
            content:
                description: The url to the post content
                type: string
                x-go-name: Content
            created_at:
                description: The time the post was created
                format: date-time
                type: string
                x-go-name: CreatedAt
            description:
                description: Post description
                type: string
                x-go-name: Description
            id:
                description: The id for this post
                format: uuid
                type: string
                x-go-name: ID
            liked_by_me:
                type: boolean
                x-go-name: LikedByMe
            likes:
                description: Number of likes
                format: int64
                type: integer
                x-go-name: Likes
            user:
                $ref: '#/definitions/User'
        required:
            - id
            - content
            - description
            - created_at
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/models
    PostCreateRequestBody:
        properties:
            content:
                type: string
                x-go-name: Content
            description:
                maxLength: 2048
                type: string
                x-go-name: Description
        required:
            - content
            - description
        title: PostCreateRequestBody includes the content, description and id of the user creating the post.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/parameters
...
    User:
        description: User represents the user for this application
        properties:
            avatar_url:
                description: URL to user avatar
                type: string
                x-go-name: AvatarURL
            created_at:
                description: The time the user was registered
                format: date-time
                type: string
                x-go-name: CreatedAt
            email:
                description: The email for this user
                type: string
                x-go-name: Email
            id:
                description: The id for this user
                format: uuid
                type: string
                x-go-name: ID
            name:
                description: The name for this user
                type: string
                x-go-name: Name
            updated_at:
                description: The time the user was updated
                format: date-time
                type: string
                x-go-name: UpdatedAt
            username:
                description: The username for this user
                type: string
                x-go-name: Username
        required:
            - id
            - email
            - username
            - created_at
            - updated_at
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/models
...
host: localhost:3040
info:
    title: Diverse API.
    version: 0.0.1
paths:
...
    /posts:
        get:
            description: Returns a list of all posts
            operationId: getPosts
            parameters:
                - format: uuid
                  in: query
                  name: last_seen_post_id
                  type: string
                  x-go-name: LastSeenPostID
                - description: nolint:lll
                  format: date-time
                  in: query
                  name: last_seen_post_created_at
                  type: string
                  x-go-name: LastSeenPostCreatedAt
                - in: query
                  name: type
                  type: string
                  x-go-name: Type
                - format: uuid
                  in: query
                  name: user_id
                  type: string
                  x-go-name: UserID
                - format: int64
                  in: query
                  maximum: 50
                  minimum: 1
                  name: count
                  required: true
                  type: integer
                  x-go-name: Count
            responses:
                "200":
                    $ref: '#/responses/GetPostsResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
        post:
            description: Creates the post with given info
            operationId: createPost
            parameters:
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/PostCreateRequestBody'
            responses:
                "201":
                    $ref: '#/responses/CreateUpdatePostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
    /posts/{post}:
        delete:
            description: Delete post by ID
            operationId: deletePost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            produces:
                - application/json
            responses:
                "204":
                    $ref: '#/responses/DeletePostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            schemes:
                - http
                - https
            security:
                - bearerAuth: []
            tags:
                - Post
        get:
            description: Returns the post by given ID
            operationId: getPost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            responses:
                "200":
                    $ref: '#/responses/GetPostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
        patch:
            description: Update post by ID with given fields
            operationId: updatePost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            responses:
                "201":
                    $ref: '#/responses/GetPostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
...
    /posts/{post}/like:
        delete:
            description: Unset like to the post by ID
            operationId: unlikePost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            responses:
                "201":
                    $ref: '#/responses/GetPostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
        post:
            description: Set like to the post by ID
            operationId: likePost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            responses:
                "201":
                    $ref: '#/responses/GetPostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
...
produces:
    - application/json
responses:
...
    CreateUpdatePostResponse:
        description: CreateUpdatePostResponse represents response for successfully create or update post request.
        schema:
            type: string
...
    DeletePostResponse:
        description: DeletePostResponse represents response for successfully delete post.
...
    ErrorResponse:
        description: |-
            ErrorResponse represents the error response body for API.
            Contains error and message fields.
        schema:
            $ref: '#/definitions/BaseResponseBody'
...
    GetPostResponse:
        description: GetPostResponse represent the response retrived on get post request.
        schema:
            $ref: '#/definitions/GetPostResponseBody'
    GetPostsResponse:
        description: GetPostsResponse represent the response retrived on get posts request.
        schema:
            $ref: '#/definitions/GetPostsResponseBody'
...
schemes:
    - http
    - https
securityDefinitions:
    bearerAuth:
        description: JWT Authorization header
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"